(modifications for #180 and #243) (1) changed variable type of 'order' and 'num_model' to unsigned int. (2) modified test/logical_operation.cc so that no warning is made in building.
git-svn-id: http://alma-dms/svn/sakura/trunk@480 9a79bd28-2a0b-46f8-beb7-14fc758f186d

+++ libsakura/src/baseline.cc
@@ -20,11 +20,11 @@ using ::Eigen::Aligned;
 namespace {
 
 inline void GetBaselineModel(
-		size_t num_chan, int order, double *out) {
+		size_t num_chan, unsigned int order, double *out) {
 	assert(LIBSAKURA_SYMBOL(IsAligned)(out));
-	size_t num_model = order + 1;
+	unsigned int num_model = order + 1;
 
-	for (size_t i = 0; i < num_model; ++i) {
+	for (unsigned int i = 0; i < num_model; ++i) {
 		for (size_t j = 0; j < num_chan; ++j) {
 			out[num_chan*i+j] = pow(static_cast<double>(j), static_cast<double>(i));
 		}
@@ -32,7 +32,7 @@ inline void GetBaselineModel(
 }
 
 inline void DoSubtractBaseline(
-		size_t num_chan, float const *in_data, bool const *in_mask, size_t num_model,
+		size_t num_chan, float const *in_data, bool const *in_mask, unsigned int num_model,
 		double const *model_data, float clipping_threshold_sigma, unsigned int num_fitting_max,
 		bool get_residual, float *out) {
 	//std::cout << "DoSubtractBaselineEigen function is called" << std::endl;
@@ -52,7 +52,7 @@ inline void DoSubtractBaseline(
 	float *residual_data = reinterpret_cast<float *>(LIBSAKURA_PREFIX::Memory::Allocate(sizeof(float)*num_chan));
 	//bool *new_clip_mask = reinterpret_cast<bool *>(LIBSAKURA_PREFIX::Memory::Allocate(sizeof(bool)*num_chan));
 
-	for (size_t i = 0; i < num_fitting_max; ++i) {
+	for (unsigned int i = 0; i < num_fitting_max; ++i) {
 		LIBSAKURA_SYMBOL(OperateLogicalAnd)(num_chan, in_mask, clip_mask, composite_mask);
 		LIBSAKURA_SYMBOL(GetBestFitModel)(num_chan, in_data, composite_mask, num_model, model_data, best_fit_model);
 		LIBSAKURA_SYMBOL(OperateFloatSubtraction)(num_chan, in_data, best_fit_model, residual_data);
@@ -98,10 +98,10 @@ inline void DoSubtractBaseline(
 
 inline void SubtractBaselinePolynomial(
 		size_t num_chan, float const *in_data, bool const *in_mask,
-		int order, float clipping_threshold_sigma,
+		unsigned int order, float clipping_threshold_sigma,
 		unsigned int num_fitting_max, bool get_residual,
 		float *out) {
-	size_t num_model = order + 1;
+	unsigned int num_model = order + 1;
 	double *model_data = reinterpret_cast<double *>(LIBSAKURA_PREFIX::Memory::Allocate(sizeof(double)*num_chan*num_model));
 
 	GetBaselineModel(num_chan, order, model_data);
@@ -116,7 +116,7 @@ inline void SubtractBaselinePolynomial(
 namespace LIBSAKURA_PREFIX {
 void ADDSUFFIX(Baseline, ARCH_SUFFIX)::SubtractBaselinePolynomial(
 		size_t num_chan, float const in_data[/*num_chan*/],
-		bool const in_mask[/*num_chan*/], int order,
+		bool const in_mask[/*num_chan*/], unsigned int order,
 		float clipping_threshold_sigma, unsigned int num_fitting_max,
 		bool get_residual, float out[/*num_chan*/]) const {
 	SubtractBaselinePolynomial(num_chan, in_data, in_mask, order,
@@ -124,13 +124,13 @@ void ADDSUFFIX(Baseline, ARCH_SUFFIX)::SubtractBaselinePolynomial(
 }
 
 void ADDSUFFIX(Baseline, ARCH_SUFFIX)::GetBaselineModel(
-		size_t num_chan, int order, double out[/*(order+1)*num_chan*/]) const {
+		size_t num_chan, unsigned int order, double out[/*(order+1)*num_chan*/]) const {
 	GetBaselineModel(num_chan, order, out);
 }
 
 void ADDSUFFIX(Baseline, ARCH_SUFFIX)::DoSubtractBaseline(
 		size_t num_chan, float const in_data[/*num_chan*/],
-		bool const in_mask[/*num_chan*/], size_t num_model,
+		bool const in_mask[/*num_chan*/], unsigned int num_model,
 		double const model_data[/*num_model * num_chan*/],
 		float clipping_threshold_sigma, unsigned int num_fitting_max,
 		bool get_residual, float out[/*num_chan*/]) const {
