+++ libsakura/src/gen_baseline.cc
@@ -8,9 +8,9 @@
 #include "libsakura/localdef.h"
 
 extern "C" LIBSAKURA_SYMBOL(Status) LIBSAKURA_SYMBOL(SubtractBaselinePolynomial)(
-		size_t num_chan, float const in_data[], bool const in_mask[], int order,
-		float clipping_threshold_sigma, unsigned int num_fitting_max,
-		bool get_residual, float out_data[]) {
+		size_t num_chan, float const in_data[], bool const in_mask[],
+		unsigned int order, float clipping_threshold_sigma,
+		unsigned int num_fitting_max, bool get_residual, float out_data[]) {
 
 	auto baselineop =
 			::LIBSAKURA_PREFIX::OptimizedImplementationFactory::GetFactory()->GetBaselineImpl();
@@ -22,7 +22,7 @@ extern "C" LIBSAKURA_SYMBOL(Status) LIBSAKURA_SYMBOL(SubtractBaselinePolynomial)
 }
 
 extern "C" LIBSAKURA_SYMBOL(Status) LIBSAKURA_SYMBOL(GetBaselineModel)(
-		size_t num_chan, int order, double out[]) {
+		size_t num_chan, unsigned int order, double out[]) {
 
 	auto baselineop =
 			::LIBSAKURA_PREFIX::OptimizedImplementationFactory::GetFactory()->GetBaselineImpl();
@@ -33,7 +33,7 @@ extern "C" LIBSAKURA_SYMBOL(Status) LIBSAKURA_SYMBOL(GetBaselineModel)(
 
 extern "C" LIBSAKURA_SYMBOL(Status) LIBSAKURA_SYMBOL(DoSubtractBaseline)(
 		size_t num_chan, float const in_data[], bool const in_mask[],
-		size_t num_model, double const model_data[], float clipping_threshold_sigma,
+		unsigned int num_model, double const model_data[], float clipping_threshold_sigma,
 		unsigned int num_fitting_max, bool get_residual, float out[]) {
 
 	auto baselineop =
