+++ libsakura/src/libsakura/optimized_implementation_factory_impl.h
@@ -39,14 +39,14 @@ public:
 	}
 	virtual void SubtractBaselinePolynomial(size_t num_chan,
 			float const in_data[/*num_chan*/],
-			bool const in_mask[/*num_chan*/], int order,
+			bool const in_mask[/*num_chan*/], unsigned int order,
 			float clipping_threshold_sigma, unsigned int num_fitting_max,
 			bool get_residual, float out[/*num_chan*/]) const;
-	virtual void GetBaselineModel(size_t num_chan, int order,
+	virtual void GetBaselineModel(size_t num_chan, unsigned int order,
 			double out[/*(order+1)*num_chan*/]) const;
 	virtual void DoSubtractBaseline(size_t num_chan,
 			float const in_data[/*num_chan*/],
-			bool const in_mask[/*num_chan*/], size_t num_model,
+			bool const in_mask[/*num_chan*/], unsigned int num_model,
 			double const model_data[/*num_model * num_chan*/],
 			float clipping_threshold_sigma, unsigned int num_fitting_max,
 			bool get_residual, float out[/*num_chan*/]) const;
@@ -58,14 +58,14 @@ public:
 	}
 	virtual void SubtractBaselinePolynomial(size_t num_chan,
 			float const in_data[/*num_chan*/],
-			bool const in_mask[/*num_chan*/], int order,
+			bool const in_mask[/*num_chan*/], unsigned int order,
 			float clipping_threshold_sigma, unsigned int num_fitting_max,
 			bool get_residual, float out[/*num_chan*/]) const;
-	virtual void GetBaselineModel(size_t num_chan, int order,
+	virtual void GetBaselineModel(size_t num_chan, unsigned int order,
 			double out[/*(order+1)*num_chan*/]) const;
 	virtual void DoSubtractBaseline(size_t num_chan,
 			float const in_data[/*num_chan*/],
-			bool const in_mask[/*num_chan*/], size_t num_model,
+			bool const in_mask[/*num_chan*/], unsigned int num_model,
 			double const model_data[/*num_model * num_chan*/],
 			float clipping_threshold_sigma, unsigned int num_fitting_max,
 			bool get_residual, float out[/*num_chan*/]) const;
